+++ docs/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 
 # v1.3.11 <small><sup>27-08-2022</sup></small>
 
+- [Docs] updated docs on isNumberOk for string parsing
 - [Docs] updated docs on isArrayOk (options.sortOrder)
 - [Docs] updated changelog